Move file-utils.ts to file-system/controllers
