+++ src/components/editor-footer.tsx
@@ -1,11 +1,14 @@
 import { AlertCircle, Download, Terminal as TerminalIcon } from "lucide-react";
 import { useUpdater } from "@/settings/hooks/use-updater";
+import {
+  getFilenameFromPath,
+  getLanguageFromFilename,
+} from "../file-system/controllers/file-utils";
 import { useFileSystemStore } from "../file-system/controllers/store";
 import { usePersistentSettingsStore } from "../settings/stores/persistent-settings-store";
 import { useBufferStore } from "../stores/buffer-store";
 import { useGitStore } from "../stores/git-store";
 import { useUIState } from "../stores/ui-state-store";
-import { getFilenameFromPath, getLanguageFromFilename } from "../utils/file-utils";
 import { getGitStatus } from "../utils/git";
 import GitBranchManager from "./git/git-branch-manager";
